refactor(config): isolate runtime timeout policy interpretation
